requirements.txt
fastf1
pandas
kafka-python
websockets